% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celsius_to_fahrenheit.R
\name{c_to_f}
\alias{c_to_f}
\title{A temperature converting Function}
\usage{
c_to_f(x)
}
\arguments{
\item{x}{A number of Celsius degree.}
}
\description{
This function allows you to convert Celsius(°C) into Fahrenheit(°F).
}
\examples{
c_to_f(19)
}
