% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fahrenheit_to_celsius.R
\name{f_to_c}
\alias{f_to_c}
\title{A temperature converting Function}
\usage{
f_to_c(x)
}
\arguments{
\item{x}{A number of Fahrenheit degree.}
}
\description{
This function allows you to convert Fahrenheit(°F) into Celsius(°C).
}
\examples{
f_to_c(70)
}
