% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goldmelt.R
\name{goldmelt}
\alias{goldmelt}
\title{A Function that tells some common physics knowledge:gold melt}
\usage{
goldmelt(x)
}
\arguments{
\item{f}{stand for fahrenheit.}

\item{c}{c stand for celsius.}
}
\description{
This function allows you to check water boiling point in Fahrenheit(F) and celsius(C).
}
\examples{
goldmelt('f')
goldmelt('c')
}
