% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ironmelt.R
\name{ironmelt}
\alias{ironmelt}
\title{A Function that tells some common physics knowledge:iron melt}
\usage{
ironmelt(x)
}
\arguments{
\item{f}{stand for fahrenheit.}

\item{c}{c stand for celsius.}
}
\description{
This function allows you to check water boiling point in Fahrenheit(??F) and celsius(??C).
}
\examples{
ironmelt('f')
ironmelt('c')
}
