% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magnesiummelt.R
\name{magnesiummelt}
\alias{magnesiummelt}
\title{A Function that tells some common physics knowledge}
\usage{
magnesiummelt(x)
}
\arguments{
\item{f}{stand for fahrenheit.}

\item{c}{c stand for celsius.}
}
\description{
This function allows you to check water boiling point in Fahrenheit(F) and celsius(C).
}
\examples{
magnesiummelt('f')
magnesiummelt('c')
}
