% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silvermelt.R
\name{silvermelt}
\alias{silvermelt}
\title{A Function that tells some common physics knowledge}
\usage{
silvermelt(x)
}
\arguments{
\item{f}{stand for fahrenheit.}

\item{c}{c stand for celsius.}
}
\description{
This function allows you to check water boiling point in Fahrenheit(F) and celsius(C).
}
\examples{
silvermelt('f')
silvermelt('c')
}
