% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zincmelt.R
\name{zincmelt}
\alias{zincmelt}
\title{A Function that tells some common physics knowledge}
\usage{
zincmelt(x)
}
\arguments{
\item{f}{stand for fahrenheit.}

\item{c}{c stand for celsius.}
}
\description{
This function allows you to check water boiling point in Fahrenheit(F) and celsius(C).
}
\examples{
zincmelt('f')
zincmelt('c')
}
